Every spend threw, and my own catch block hid it
`SPEND_LOG_SQL` read `-$2` and Postgres refused it every single time:

    42725  operator is not unique: - unknown

A bare parameter under unary minus has no inferable type, so the statement
is ambiguous before it ever looks at a value. `-($2::int)` fixes it.

Verified against production rather than reasoned about: two real turns
moved the QA balance 2000 -> 1999 -> 1998, and the ledger held nothing but
the two original grants. Then the statement run directly against the same
database, which named the error in one line.

WHY IT WAS INVISIBLE is the part worth keeping. The caller catches this
insert and does not rethrow — which is correct, because the credits have
already gone and a broken audit table must not take down a turn whose
balance has already moved. But a catch that exists to protect a turn will
just as happily hide a statement that has never once succeeded. I wrote
that catch, wrote a comment about it failing being "the safe direction",
and never asked whether it was failing.

GRANT_LOG_SQL was fine throughout: its amount is positive and needs no
operator to type it. That asymmetry is exactly why grants showed up in the
ledger and spends did not, and it is what made the table look like it was
working.

Two bugs, both real, both needed. The previous commit awaited the insert —
without that a working statement still would not land, because the pooled
client is released the instant the function returns. This one makes the
statement work at all.

+++ apps/api/src/decke/credits.ts
@@ -153,10 +153,28 @@ export const SPEND_SQL = `
  * If this fails the credits are still gone, which is the safe direction: a
  * missing log line is a gap in a statement, where a missing decrement is free
  * work. The log is the evidence and the balance is the product.
+ *
+ * ── `-($2::int)`, AND THE CAST IS NOT DECORATION ─────────────────────────────
+ *
+ * This read `-$2` and threw on every single call:
+ *
+ *     42725  operator is not unique: - unknown
+ *
+ * Postgres cannot infer a type for a bare parameter under unary minus, so the
+ * statement is ambiguous before it ever looks at the value. Verified against
+ * production: the balance moved 2000 → 1999 → 1998 across two real turns and the
+ * ledger held nothing but the two original grants.
+ *
+ * It was invisible because the caller catches this insert and does not rethrow —
+ * correctly, since the credits are already gone and a broken audit table must
+ * not take down a turn. A `.catch()` that exists to protect a turn will also
+ * hide a statement that has never once succeeded. `GRANT_LOG_SQL` was fine
+ * throughout because its `$2` is positive and needs no operator to type it,
+ * which is why grants appeared in the ledger and spends did not.
  */
 export const SPEND_LOG_SQL = `
   INSERT INTO decke_credit_event (user_id, delta, kind, reason)
-  VALUES ($1, -$2, 'spend', $3)`;
+  VALUES ($1, -($2::int), 'spend', $3)`;
 
 /**
  * A grant. Parameters: `$1` = user id, `$2` = credits, `$3` = reason,
